scripts/auth/requirements.txt
kurigram
tgcrypto
python-dotenv
